==> analysis/Notebook 5 - SHet SHom Tests.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from qmplot import manhattanplot
import re
from numpy.linalg import inv
from pathlib import Path 
from scipy.stats import chisquare
import scipy.stats as stats  

df_candidate = pd.read_table("../output_csv/df_candidate.csv", low_memory=False, sep=";")

df_candidate

z = df_candidate[['fi_z', 'fg_z', 'hba1c_z', 't2d_z', 'bmi_z', 'gcn_z']].to_numpy()

z

z.shape

correlation = np.loadtxt('correlation.txt', usecols=range(6))

correlation

fg_n = 200622
fi_n = 151013
hba1c_n = 146806
bmi_n = 806834
t2d_n = 492191
gcn_n = 8421

w_i = [fi_n, fg_n, hba1c_n, t2d_n, bmi_n, gcn_n]

sum_wi = np.sqrt(np.sum(np.square(w_i)))

w = w_i/sum_wi

w

sign_t = np.sign(z)

sign_t = np.nan_to_num(sign_t)
sign_t

w_s = w * sign_t

w_s

w_s[0]

correlation

inv(correlation)

sig = np.linalg.pinv(correlation)

sig

z = np.nan_to_num(z)

z

w = w.reshape(1,6)
w

z.shape

sHom_scores = []
count = 1
sig = inv(correlation)
for i in range(z.shape[0]):
    k = np.matmul(np.matmul(w, sig), np.asmatrix(z[i]).T)
    p = np.matmul(np.matmul(w, sig), np.asmatrix(w).T)
    s = (k*k)/p
    score = np.asarray(s).flatten()[0]
    sHom_scores.append(score)
    if count % 1000000 == 0:
        print("-----", count, " iterations -------")
    count+= 1

np.savetxt('sHom_scores.txt', sHom_scores)

sHom_scores = np.loadtxt('sHom_scores.txt', usecols=range(1))
from scipy.stats import chisquare
from rpy2.robjects import r, pandas2ri

pandas2ri.activate()
p_values = []
count = 0
for s in sHom_scores:
    if(count %1000000 == 0):
        print("------", count, "----------")
    p = np.exp(np.longdouble(r.pchisq(s,1, lower_tail=False, log_p=True)[0]))
    p_values.append(p)
    #6 degree of freedom
    #p_value = stats.chi2.cdf(scores, 6)
    #p_values.append(p_value) 
    count += 1

np.savetxt('sHom_p.txt', p_values)

p_values

from scipy.stats import chi2
count = 0
sHet_scores = []
cut_off = 0
for i in range(z.shape[0]): 
    T = -1
    true_list, = np.where(np.abs(z[i]) > cut_off)
    w_i = w
    t = z[i]
    cor_i = np.asarray(correlation)
    index = -1
    snp_size = z.shape[1]
    for j in range(z.shape[1]):
        index += 1
        if j not in true_list:
            cor_i = np.delete(cor_i, index, 1)
            cor_i = np.delete(cor_i, index, 0)
            w_i = np.delete(w_i, index)
            snp_size -= 1
            w_i = w_i.reshape(1, snp_size)
            t = np.delete(t, index)
            index -= 1
        else:
            if(t[index] < 0):
                w_i[0][index] = - w_i[0][index]
    sig_i = np.linalg.pinv(cor_i) 
    k = np.matmul(np.matmul(w_i, sig_i),np.asmatrix(t).T)
    p = np.matmul(np.matmul(w_i, sig_i),np.asmatrix(w_i).T)
    s = (k * k)/p
    score = s.flatten().tolist()[0][0]
    if(score > T):
        sHet_scores.append(score)
    else:
        sHet_scores.append(T)
    if count % 1000000 == 0:
        print("-----", count, " iterations -------")
    count+= 1

sHet_scores

np.savetxt('SHet_scores.txt', sHet_scores)

fit_alpha, fit_loc, fit_beta=stats.gamma.fit(sHet_scores)
print(fit_alpha, fit_loc, fit_beta)

p_zero = stats.gamma.pdf(0, fit_alpha, loc=fit_loc, scale=fit_beta)
SHet_p = []
count = 0
for s in sHet_scores:
    if count % 1000000 == 0:
        print("-----", count, " iterations -------")
    if(s == 0):
        SHet_p.append(p_zero)
    else:
        SHet_p.append(stats.gamma.pdf(s, fit_alpha, loc=fit_loc, scale=fit_beta))
    count+= 1

np.savetxt('SHet_p.txt', SHet_p)
