==> analysis/Notebook 6 - Combine scores.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from qmplot import manhattanplot
import re
from numpy.linalg import inv
from pathlib import Path 

Omni_z =np.loadtxt('omni_z.txt', usecols=range(1))

df = pd.read_table("../output_csv/df_candidate.csv", low_memory=False, sep=";")

def formatdf_plot(df, p_values):
    df = df.rename(columns={"chromosome": "CHR", "base_pair_location": "BP"})
    df["P"] = p_values
    df.dropna(subset=['CHR', 'P'])
    
    df["SNP"] = df["BP"]
    df["GENE"] = "Not Given"
    df["DISTANCE"] = "Not Given" 
    
    df = df.sort_values(by = ['CHR', "BP"]).reset_index()

    return df

df

df_omni_z = formatdf_plot(df, Omni_z)

df_omni_z

df_omni_plot = df_omni_z.dropna()

df_omni_plot = df_omni_plot.drop(columns='index').reset_index().drop(columns="index")

df_omni_plot[df_omni_plot.P < 1e-6].reset_index().drop(columns="index")

import dash_bio
dash_bio.ManhattanPlot(
    dataframe=df_omni_plot[df_omni_plot.P < 1e-6].reset_index().drop(columns="index"),
    highlight_color='#00FFAA',
    suggestiveline_color='#AA00AA',
    genomewideline_color='#AA5500'
)

def plot_manhatten_plots(df, figname):
    # generate manhattan plot and set an output file.
    colors = plt.cm.PuBuGn(np.linspace(0, 0.7, 10))
    xtick = set(['chr' + i for i in list(map(str, range(1, 10))) + ['11', '13', '15', '18', '21', '22', 'X', 'XY', 'MT']])
    f, ax = plt.subplots(figsize=(12, 4), facecolor='w', edgecolor='k')
    manhattanplot(data=df,
                  marker=".",
                  sign_marker_p=5e-8,  # Genome wide significant p-value
                  sign_marker_color=colors[7],
                  snp="POS",

                  title=figname,
                  xlabel="Chromosome",
                  ylabel=r"$-log_{10}{(P)}$",

                  sign_line_cols=["#D62728", "#2CA02C"],
                  hline_kws={"linestyle": "--", "lw": 1.3},
                  ld_block_size=500000,  # 500000 bp
                  text_kws={"fontsize": 12,  # The fontsize of annotate text
                            "arrowprops": dict(arrowstyle="-", color="k", alpha=0.6)},
                  ax=ax)
    plt.title(figname)
    plt.savefig(f'../output_figs/{re.sub(r"[^a-zA-Z0-9]","",figname)}.jpg', dpi=400, bbox_inches='tight')
    plt.show()

df_omni_z[df_omni_z.P < 1e-15]

df_omniz_plot = df_omni_z.dropna()

df_omniz_plot[df_omniz_plot.P < 1e-8]

df_omniz_plot.loc[df_omniz_plot.P == 0, 'P'] = 1e-323

plot_manhatten_plots(df_omniz_plot[df_omniz_plot.P < 1e-8], "Omni Z")

Sum_z =np.loadtxt('sum_z.txt', usecols=range(1))

df_sum_z = formatdf_plot(df, Sum_z)

df_sum_z

import dash_bio
dash_bio.ManhattanPlot(
    dataframe=df_sum_z[df_sum_z.P < 1e-6].reset_index().drop(columns="index"),
    highlight_color='#00FFAA',
    suggestiveline_color='#AA00AA',
    genomewideline_color='#AA5500',
    title="Sumz Test Manhattan Plot"
)

df_sum_z_plot = df_sum_z.dropna(subset=['P'])
df_sum_z_plot

df_sum_z_plot[df_sum_z_plot.P < 5e-3]

plot_manhatten_plots(df_sum_z_plot[df_sum_z_plot.P < 5e-3], "Sum Z")

SHom_scores = np.loadtxt('sHom_p.txt', usecols=range(1))

df_SHom_scores = formatdf_plot(df, SHom_scores)

df_SHom_scores

df_SHom_scores[df_SHom_scores.P < 1e-6].reset_index().drop(columns="index")

dash_bio.ManhattanPlot(
    dataframe=df_SHom_scores[df_SHom_scores.P < 1e-6].reset_index().drop(columns="index"),
    highlight_color='#00FFAA',
    suggestiveline_color='#AA00AA',
    genomewideline_color='#AA5500',
    title = "SHom Test - Manhattan Plot"
)

df_SHom_scores = df_SHom_scores.dropna(subset='P')
df_SHom_scores

plot_df = df_SHom_scores[(np.abs(df_SHom_scores.P) >0) & (np.abs(df_SHom_scores.P) < 1e-6) ]

plot_df.P = np.abs(plot_df.P)

plot_manhatten_plots(plot_df, "SHom Test")

import matplotlib.pyplot as plt
from qmplot import qqplot

# Create a Q-Q plot
f, ax = plt.subplots(figsize=(6, 6), facecolor="w", edgecolor="k")
qqplot(data=plot_df[plot_df.P < 1e-6].P,
       marker="o",
       title="SHom P",
       xlabel=r"Expected $-log_{10}{(P)}$",
       ylabel=r"Observed $-log_{10}{(P)}$",
       ax=ax)

plt.savefig("../output_figs/sHom.QQ.png")

SHet_scores = np.loadtxt('SHet_p.txt', usecols=range(1))

df_SHet_scores = formatdf_plot(df, SHet_scores)

df_SHet_scores[df_SHet_scores.P < 1e-8]

dash_bio.ManhattanPlot(
    dataframe=df_SHet_scores[df_SHet_scores.P < 1e-8].reset_index().drop(columns="index"),
    highlight_color='#00FFAA',
    suggestiveline_color='#AA00AA',
    genomewideline_color='#AA5500',
    title = "SHet Test - Manhattan Plot"
)

from scipy import stats
import plotly.graph_objects as go
import plotly.graph_objects as go
qq = stats.probplot(df_SHet_scores[df_SHet_scores.P<1e-6].reset_index().drop(columns='index').P, dist='lognorm', sparams=(1))
x = np.array([qq[0][0][0], qq[0][0][-1]])

fig = go.Figure()
fig.add_scatter(x=qq[0][0], y=qq[0][1], mode='markers')
fig.add_scatter(x=x, y=qq[1][1] + qq[1][0]*x, mode='lines')
fig.layout.update(showlegend=False)
fig.show()

qq = stats.probplot(df_SHom_scores[df_SHom_scores.P<1e-6].reset_index().drop(columns="index").P, dist='lognorm', sparams=(1))
x = np.array([qq[0][0][0], qq[0][0][-1]])
fig = go.Figure()
fig.add_scatter(x=qq[0][0], y=qq[0][1], mode='markers')
fig.add_scatter(x=x, y=qq[1][1] + qq[1][0]*x, mode='lines')
fig.layout.update(showlegend=False)
fig.show()

plot_manhatten_plots(df_SHet_scores[(df_SHet_scores.P > 0) & (df_SHet_scores.P < 1e-5)], "SHet Test")

# Create a Q-Q plot
f, ax = plt.subplots(figsize=(6, 6), facecolor="w", edgecolor="k")
qqplot(data=df_SHet_scores[(df_SHet_scores.P > 0) & (df_SHet_scores.P < 1e-5)].P,
       marker="o",
       title="SHet P",
       xlabel=r"Expected $-log_{10}{(P)}$",
       ylabel=r"Observed $-log_{10}{(P)}$",
       ax=ax)

plt.savefig("sHet.QQ.png")

def plot_significant_loci_counts(df, dataset_name):
    df = df.rename(columns={"#CHROM": "chromosome", "POS": "base_pair_location", "P": "p_value"})
    list_chromosomes = []
    for chromosome in set(df.chromosome):
        list_loci = []
        bp_pos = []
        df_chromosome = df[df.chromosome == chromosome]
        df_chromosome = df_chromosome.sort_values(by = ['base_pair_location']).reset_index()
        df_chromosome = df_chromosome.drop(['index'], axis=1)
        df_chromosome.reset_index(inplace=True, drop=True); 
        df_significant_chromosome = df_chromosome[np.abs(df_chromosome['p_value']) < 5e-8]
        print(df_significant_chromosome.shape[0])
        if(len(df_significant_chromosome) > 0):
            first_bp = df_significant_chromosome.base_pair_location.to_list()[0]
            counter = 1
            for bp in df_significant_chromosome.base_pair_location:
                if first_bp == 0:
                    bp_pos = []
                    bp_pos.append(first_bp)
                    first_bp = bp
                    counter = 0
                elif(bp - first_bp < 1000000):
                    counter+= 1
                    bp_pos.append(bp)
                    continue
                else:
                   #if(counter > 10): 
                   list_loci.append({"start_pos": first_bp, "end_pos": bp_pos[len(bp_pos) - 1], "bp_pos":bp_pos, "sig_count": counter})
                   first_bp = 0
            list_chromosomes.append({'chromosome': chromosome, 'list_significant_loci': list_loci})
    x = []
    y = []
    for row in list_chromosomes:
        chrom = row['chromosome']
        chrom_int = 0
        if(chrom == "X"):
            chrom_int = 23
        elif chrom == "XY": 
            chrom_int = 24
        elif chrom == "MT":
            chrom_int = 25
        elif chrom == "Y":
            chrom_int = 26
        else:
            chrom_int = chrom
        #we can plot with only int values in bar chart hence encoded to int values
        x.append(chrom_int)
        y.append(len(row['list_significant_loci']))

    
    ##fig = plt.figure()
    #ax = fig.add_axes([0,0,1,1])
    #ax.bar(x,y)
    #ax.set_title(f'{dataset_name}: Significant Loci Count per chromosome')
    #plt.xticks(rotation='vertical', fontsize = 5)
    #plt.gca().xaxis.set_major_locator(plt.MultipleLocator(1))
    #plt.savefig(f'../output_figs/{re.sub(r"[^a-zA-Z0-9]","",dataset_name)}_significant_loci_graph.jpg', dpi=400, bbox_inches='tight')
    #plt.show()
    #print(sum(y))
    total_sig = sum(y)
    return list_chromosomes, total_sig


df_omni_z

def count_suggestive_loci_counts(df, dataset_name, sig):
    df = df.rename(columns={"CHR": "chromosome", "BP": "base_pair_location", "P": "p_value"})
    list_chromosomes = []
    for chromosome in set(df.chromosome):
        list_loci = []
        bp_pos = []
        df_chromosome = df[df.chromosome == chromosome]
        df_chromosome = df_chromosome.sort_values(by = ['base_pair_location']).reset_index()
        df_chromosome = df_chromosome.drop(['index'], axis=1)
        df_chromosome.reset_index(inplace=True, drop=True); 
        df_significant_chromosome = df_chromosome[df_chromosome['p_value'] < sig]
        if(len(df_significant_chromosome) > 0):
            first_bp = 0 
            counter = 1
            for bp in df_significant_chromosome.base_pair_location:
                if first_bp == 0:
                    bp_pos = []
                    bp_pos.append(first_bp)
                    first_bp = bp
                    counter = 0
                elif(bp - first_bp < 1000000):
                    counter+= 1
                    bp_pos.append(bp)
                    continue
                else:
                   #if(counter > 10): 
                   list_loci.append({"start_pos": first_bp, "end_pos": bp_pos[len(bp_pos) - 1], "bp_pos":bp_pos, "sig_count": counter})
                   first_bp = 0
            list_loci.append({"start_pos": first_bp, "end_pos": bp_pos[len(bp_pos) - 1], "bp_pos":bp_pos, "sig_count": counter})
            list_chromosomes.append({'chromosome': chromosome, 'list_significant_loci': list_loci})
    x = []
    y = []
    for row in list_chromosomes:
        chrom = row['chromosome']
        chrom_int = 0
        if(chrom == "X"):
            chrom_int = 23
        elif chrom == "XY": 
            chrom_int = 24
        elif chrom == "MT":
            chrom_int = 25
        elif chrom == "Y":
            chrom_int = 26
        else:
            chrom_int = chrom
        #we can plot with only int values in bar chart hence encoded to int values
        x.append(chrom_int)
        y.append(len(row['list_significant_loci']))

    total_sig = sum(y)
    return list_chromosomes, total_sig

omni_sig_loci, omni_total_sig = count_suggestive_loci_counts(df_omni_z, "Omni Z", 5e-8)

filepath = Path('../output_csv/df_omni_z.csv')  
filepath.parent.mkdir(parents=True, exist_ok=True) 
#save progress after omni p values
df_omni_z.to_csv(filepath, index=False, sep = ';') 

omni_total_sig

sumz_sig_loci, sumz_total_sig = count_suggestive_loci_counts(df_sum_z, "Sum Z", 5e-8)
sumz_total_sig

filepath = Path('../output_csv/df_sum_z.csv')  
filepath.parent.mkdir(parents=True, exist_ok=True) 
#save progress after omni p values
df_sum_z.to_csv(filepath, index=False, sep = ';') 

shet_sig_loci, shet_total_sig = count_suggestive_loci_counts(df_SHet_scores, "SHet", 5e-8)
shet_total_sig

filepath = Path('../output_csv/df_SHet_scores.csv')  
filepath.parent.mkdir(parents=True, exist_ok=True) 
#save progress after omni p values
df_SHet_scores.to_csv(filepath, index=False, sep = ';') 

shom_sig_loci, shom_total_sig = count_suggestive_loci_counts(df_SHom_scores, "SHom", 5e-8)
shom_total_sig

filepath = Path('../output_csv/df_SHom_scores.csv')  
filepath.parent.mkdir(parents=True, exist_ok=True) 
#save progress after omni p values
df_SHom_scores.to_csv(filepath, index=False, sep = ';') 

x = ["Omni Z", "Sum Z", "SHet", "SHom"]
y = [omni_total_sig, sumz_total_sig, shet_total_sig, shom_total_sig]

fig = plt.figure()
ax = fig.add_axes([0,0,1,1])
ax.bar(x,y, color="#04D8B2")
ax.set_title('Significant Loci Count per test')
plt.xticks(rotation='vertical', fontsize = 8)
plt.gca().xaxis.set_major_locator(plt.MultipleLocator(1))
plt.savefig(f'../output_figs/{re.sub(r"[^a-zA-Z0-9]","","Significant Loci Count per test")}.jpg', dpi=400, bbox_inches='tight')
plt.show()

df_SHet_scores[np.abs(df_SHet_scores['P']) < 5e-8]

df_candidate = pd.read_table("../output_csv/df_candidate.csv", low_memory=False, sep=";")

df_candidate

df_sig = df_candidate[df_candidate.]

df_omni_sig = df_omni_z[df_omni_z.P < 5e-8]

df_omni_sig

df_omni_sig[:][['CHR', 'BP']].to_numpy()

noval_loci = []
for chrom, pos, p in df_omni_sig[:][['CHR', 'BP', 'P']].to_numpy():
    found = False
    for chrom_check, pos_check in df_sig[:][['chromosome', 'base_pair_location']].to_numpy():
        if(chrom == chrom_check) and (pos == pos_check):
            found = True
    if(found == False):
        noval_loci.append([chrom, pos, p])
noval_loci
        
            

column_values = ['chromosome', 'base_pair_location', 'p_value']
  
# creating the dataframe
df_omni_noval_loci = pd.DataFrame(data = noval_loci, 
                  columns = column_values)
  
# displaying the dataframe
df_omni_noval_loci

omni_noval_loci, omni_total_noval = count_suggestive_loci_counts(df_omni_noval_loci, "Omni Z", 5e-8)

omni_total_noval

df_sumz_sig = df_sum_z[df_sum_z.P < 5e-8]

sumz_noval_loci = []
for chrom, pos, p in df_sumz_sig[:][['#CHROM', 'POS', 'P']].to_numpy():
    found = False
    for chrom_check, pos_check in df_sig[:][['chromosome', 'base_pair_location']].to_numpy():
        if(chrom == chrom_check) and (pos == pos_check):
            found = True
    if(found == False):
        sumz_noval_loci.append([chrom, pos, p])
sumz_noval_loci

column_values = ['chromosome', 'base_pair_location', 'p_value']
  
# creating the dataframe
df_sumz_noval_loci = pd.DataFrame(data = sumz_noval_loci, 
                  columns = column_values)
  
# displaying the dataframe
df_sumz_noval_loci

sumz_noval_loci, sumz_total_noval = count_suggestive_loci_counts(df_sumz_noval_loci, "Sum Z", 5e-8)

sumz_total_noval

df_SHet_sig = df_SHet_scores[np.abs(df_SHet_scores.P) < 5e-8]

sHet_noval_loci = []
for chrom, pos, p in df_SHet_sig[:][['#CHROM', 'POS', 'P']].to_numpy():
    found = False
    for chrom_check, pos_check in df_sig[:][['chromosome', 'base_pair_location']].to_numpy():
        if(chrom == chrom_check) and (pos == pos_check):
            found = True
    if(found == False):
        sHet_noval_loci.append([chrom, pos, p])
sHet_noval_loci

column_values = ['chromosome', 'base_pair_location', 'p_value']
  
# creating the dataframe
df_sHet_noval_loci = pd.DataFrame(data = sHet_noval_loci, 
                  columns = column_values)
  
# displaying the dataframe
df_sHet_noval_loci

sHet_noval_loci, sHet_total_noval = count_suggestive_loci_counts(df_sHet_noval_loci, "SHet", 5e-8)

sHet_total_noval

df_SHom_sig = plot_df[np.abs(plot_df.P) < 5e-8]

df_SHom_sig

sHom_noval_loci = []
for chrom, pos, p in df_SHom_sig[:][['#CHROM', 'POS', 'P']].to_numpy():
    found = False
    for chrom_check, pos_check in df_sig[:][['chromosome', 'base_pair_location']].to_numpy():
        if(chrom == chrom_check) and (pos == pos_check):
            found = True
    if(found == False):
        sHom_noval_loci.append([chrom, pos, p])
sHom_noval_loci

column_values = ['chromosome', 'base_pair_location', 'p_value']
  
# creating the dataframe
df_sHom_noval_loci = pd.DataFrame(data = sHom_noval_loci, 
                  columns = column_values)
  
# displaying the dataframe
df_sHom_noval_loci

sHom_noval_loci, sHom_total_noval = count_suggestive_loci_counts(df_sHom_noval_loci, "SHom", 5e-8)

sHom_total_noval

x = ["Omni Z", "Sum Z", "SHet", "SHom"]
y = [omni_total_noval, sumz_total_noval, sHet_total_noval, sHom_total_noval]

fig = plt.figure()
ax = fig.add_axes([0,0,1,1])
ax.bar(x,y)
ax.set_title('Significant Loci Count per test')
plt.xticks(rotation='vertical', fontsize = 8)
plt.gca().xaxis.set_major_locator(plt.MultipleLocator(1))
plt.savefig(f'../output_figs/{re.sub(r"[^a-zA-Z0-9]","","Noval Loci Count per test")}.jpg', dpi=400, bbox_inches='tight')
plt.show()

df_SHom_scores

from scipy.stats import chi2
p_value = chi2.pdf(df_SHom_scores.P, 1)

p_value

df_SHom_scores['chisq_p'] = p_value

df_SHom_scores[(np.abs(df_SHom_scores.chisq_p) > 0) & (np.abs(df_SHom_scores.chisq_p) < 1e-6)]
